requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
aiofiles==24.1.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   -r requirements.in
    #   discord-py
    #   wom-py
aiomysql==0.2.0
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
alembic==1.16.5
    # via -r requirements.in
apscheduler==3.11.0
    # via -r requirements.in
attrs==25.3.0
    # via aiohttp
audioop-lts==0.2.2
    # via
    #   -r requirements.in
    #   discord-py
black==25.9.0
    # via -r requirements.in
cachetools==6.2.0
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via black
concurrent-log-handler==0.9.28
    # via -r requirements.in
discord-py==2.6.3
    # via
    #   -r requirements.in
    #   reactionmenu
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.41.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.2
    # via gspread
greenlet==3.2.4
    # via sqlalchemy
gspread==6.2.1
    # via -r requirements.in
httplib2==0.31.0
    # via oauth2client
idna==3.10
    # via
    #   requests
    #   yarl
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
msgspec==0.19.0
    # via wom-py
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via black
pathspec==0.12.1
    # via black
pillow==11.3.0
    # via -r requirements.in
platformdirs==4.4.0
    # via black
portalocker==3.2.0
    # via concurrent-log-handler
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via
    #   google-auth
    #   oauth2client
pymysql==1.1.2
    # via aiomysql
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
pytokens==0.1.10
    # via black
pytz==2025.2
    # via -r requirements.in
reactionmenu==3.1.7
    # via -r requirements.in
requests==2.32.5
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via
    #   google-auth
    #   oauth2client
six==1.17.0
    # via
    #   oauth2client
    #   python-dateutil
sqlalchemy[asyncio]==2.0.43
    # via
    #   -r requirements.in
    #   alembic
strenum==0.4.15
    # via -r requirements.in
tabulate==0.9.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   alembic
    #   sqlalchemy
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via requests
wom-py==2.0.3
    # via -r requirements.in
yarl==1.20.1
    # via aiohttp
